도서 검색 및 리스트 출력 구현완료

+++ src/main/java/com/example/app/mapper/BoardMapper.java
@@ -18,7 +18,7 @@ public interface BoardMapper {
     // 게시글 추가
     public void insert(BoardDTO boardDTO);
     // 게시글 삭제
-    public void delete(int anId);
+    public void delete(Long anId);
     // 게시글 수정
     public void update(BoardDTO boardDTO);
     // 전체 게시글 갯수
